+++ image/scripts/dev/docker-registry/all_node_registry.sh
@@ -5,28 +5,10 @@ set -o pipefail
 
 # This file executes settings and installations per K8s node.
 # This file must run as root.
-export REGISTRY_ETC_DIR=/etc/rancher/k3s
-export DOCKER_REGISTRY_DIR=/vagrant/image/scripts/dev/docker-registry
-export REGISTRY_TEMP_DIR=/vagrant/tmp
 export ETC_HOSTS=/etc/hosts
 DOCKER_REGISTRY_PORT=30099
 FQDN="${1}"
 HOST_IP="${2}"
-
-function registerRegistryWithK3s() {
-  local fqdnTemplateVar="REGISTRY_IDENTIFIER"
-  local registryTemplate="${DOCKER_REGISTRY_DIR}/k3s_registry.tmpl.yaml"
-  local registryTemplateRendered="${REGISTRY_TEMP_DIR}/k3s_registry.yaml"
-  local registryTargetPath="${REGISTRY_ETC_DIR}/registries.yaml"
-
-  mkdir -p "${REGISTRY_TEMP_DIR}"
-  mkdir -p "${REGISTRY_ETC_DIR}"
-
-  echo "Registering Registry with K3s"
-  sed "s/${fqdnTemplateVar}/${FQDN}:${DOCKER_REGISTRY_PORT}/g" "${registryTemplate}" > "${registryTemplateRendered}"
-
-  cp "${registryTemplateRendered}" "${registryTargetPath}"
-}
 
 function addEtcHostsEntryForRegistry() {
   local etcHostsExitCode=0
@@ -53,16 +35,9 @@ function echoPushHint() {
 "
 }
 
-function restartK3s() {
-  echo "Restarting k3s to apply registry.yaml..."
-  systemctl restart k3s || systemctl restart k3s-agent
-}
-
 function runSetNode() {
-  registerRegistryWithK3s
   addEtcHostsEntryForRegistry
   echoPushHint
-  restartK3s
 }
 
 # make the script only run when executed, not when sourced from bats tests)